requirements.txt
maturin
Pillow
numpy